Implement dispatchKeyEventToSession()
Now the key codes registered in the whitelist are properly passed to the
underlying TV input session.

Also changed the key code to open the menu from DPAD_LEFT/RIGHT to
DPAD_CENTER because it is the key that most people naturally press with
Bemote when they first use the TV app.

Change-Id: Ib3b557ee8c04955682002b760cc5c09ae3f8357f