Don't add BUILD_NUMBER to version for platform builds
We're removing differences between two builds that have the same source.
These builds will have different BUILD_NUMBERs though, so it shouldn't
be embedded in the binary.

For eng builds, using the current date forces the APK to be rebuilt
every time the makefiles are read. Remove this for platform builds, but
keep it if building unbundled.

Bug: 24201956
Test: Inspect ninja file for changes between two runs
Change-Id: Ib42efb54ae6c6859389eb5e81b6a6f337696d049